First commit of banking app
Set up the banking up, which is composed of 4 modules:
- bank-api
- bank-domain
- bank-dal
- bank-application

In this commit you will also find some Unit tests,
some integration tests and a basic dockerisation for the app.

I wanted to make incremental commits but I had an
inconvenience .. next time for sure! hopefully
for the future features/bug fixes to come!

In order to test the app spin it up
using the readme and test using swagger api! cheers