Add security to bank application
Add a new project to handle the security of the app. This module will provide a base
for enabling web security. The securing will be done on two axis:
- Authenticating each user: when this operation is successful it yields a valid token that expires in a day.
- Authorizing the user to a resource based on his Role/ authorities: the deposit/withdraw operations are
available for everyone. The create and read accounts is available for managers (manager/password) and admins (admin/password).
The delete account operation in the other hand requires a manager role as it's the one contain account::delete permission.

PS: In order to have the above users registered you need to add the 'dev' profile to the database profile.

Removing a trailing unused configuration class that was no longer needed