Implement BI events and kafka messaging for bank
When a new operation is successful dispatch a BI Event.
This event will be intercepted in a consumer via Application event listener.
Once the event queue is filled, we trigger the publish method
that will hand the event to all subscribers.
Subscribers then can decide what to do with the given events.

Add some async message in the new kafka module
Contains a Kafka producer that will publish the BI events that happened in the application when he receives them from the Bi
event publisher.
A consumer is also listening to the topic of the Kafka producer and will consume these events. To showcase of the retry mechanism in case
a read operation is not successful, hence not committed, the Dummy Notification Service
will throw an error when the consumed event has an odd account id!
The kafka implementation uses Avro and schema registry in order to control the quality of the data stored and insure a compatibility
between the versions of the Operation's schema.